# api/routes/session/session.route.js
"use strict";
var __create = Object.create;
var __defProp = Object.defineProperty;
var __getOwnPropDesc = Object.getOwnPropertyDescriptor;
var __getOwnPropNames = Object.getOwnPropertyNames;
var __getProtoOf = Object.getPrototypeOf;
var __hasOwnProp = Object.prototype.hasOwnProperty;
var __export = (target, all) => {
  for (var name in all)
    __defProp(target, name, { get: all[name], enumerable: true });
};
var __copyProps = (to, from, except, desc) => {
  if (from && typeof from === "object" || typeof from === "function") {
    for (let key of __getOwnPropNames(from))
      if (!__hasOwnProp.call(to, key) && key !== except)
        __defProp(to, key, { get: () => from[key], enumerable: !(desc = __getOwnPropDesc(from, key)) || desc.enumerable });
  }
  return to;
};
var __toESM = (mod, isNodeMode, target) => (target = mod != null ? __create(__getProtoOf(mod)) : {}, __copyProps(
  // If the importer is in node compatibility mode or this is not an ESM
  // file that has been converted to a CommonJS file using a Babel-
  // compatible transform (i.e. "__esModule" has not been set), then set
  // "default" to the CommonJS "module.exports" for node compatibility.
  isNodeMode || !mod || !mod.__esModule ? __defProp(target, "default", { value: mod, enumerable: true }) : target,
  mod
));
var __toCommonJS = (mod) => __copyProps(__defProp({}, "__esModule", { value: true }), mod);

// app/routes/session/session.route.ts
var session_route_exports = {};
__export(session_route_exports, {
  default: () => session_route_default
});
module.exports = __toCommonJS(session_route_exports);
var import_express = require("express");

// middlewares/authenticate-request.ts
var import_dotenv = __toESM(require("dotenv"));
var import_jsonwebtoken = __toESM(require("jsonwebtoken"));

// models/user.model.ts
var import_mongoose = require("mongoose");
var userSchema = new import_mongoose.Schema(
  {
    first_name: { type: String, required: true },
    last_name: { type: String, required: true },
    email: { type: String, required: true },
    password: { type: String, required: true },
    pass_salt: String,
    verification_token: String,
    valid_for: Date,
    isVerified: Boolean
  },
  { timestamps: true }
);
var User = (0, import_mongoose.model)("User", userSchema);

// middlewares/authenticate-request.ts
import_dotenv.default.config();
var authenticateToken = async (req, res, next) => {
  try {
    const authHeader = req.headers["authorization"];
    const token = authHeader?.startsWith("Bearer ") ? authHeader.split(" ")[1] : null;
    if (!token) {
      res.status(401).jsend.fail({ code: 401, message: "Missing Bearer token" });
      return;
    }
    const secret = process.env.ACCESS_SECRET;
    if (!secret) throw new Error("JWT secret not configured");
    const decoded = import_jsonwebtoken.default.verify(token, secret);
    const user = await User.findById(decoded.id);
    if (!user) {
      res.status(401).jsend.fail({ code: 401, message: "User not found or deleted" });
      return;
    }
    const {
      verification_token,
      isVerified,
      valid_for,
      password,
      pass_salt,
      ...other
    } = user.toObject();
    const userInReq = { ...other };
    req.user = userInReq;
    next();
  } catch (err) {
    console.error("Auth error:", err);
    res.status(401).jsend.fail({ code: 401, message: "Invalid or expired token" });
    return;
  }
};

// models/session.model.ts
var import_mongoose2 = require("mongoose");
var sessionSchema = new import_mongoose2.Schema(
  {
    user: { type: import_mongoose2.Schema.Types.ObjectId, ref: "User", required: true },
    refresh: { type: String, required: true },
    device_name: { type: String, required: true },
    device_id: { type: String, required: true },
    ip_add: { type: String, required: true },
    ip_location: { type: String, required: true },
    latitude: { type: String, required: true },
    longitude: { type: String, required: true }
  },
  { timestamps: true }
);
var Session = (0, import_mongoose2.model)("Session", sessionSchema);

// app/routes/session/session.service.ts
var import_mongoose3 = require("mongoose");
var SessionService = {
  async getUserSessions(userId, res) {
    try {
      if (!import_mongoose3.Types.ObjectId.isValid(userId)) {
        return res.status(400).jsend.fail({ code: 400, message: "Invalid user ID" });
      }
      const sessions = await Session.find({
        user: userId
      });
      res.status(200).jsend.success({
        code: 200,
        message: "Sessions are ready",
        data: sessions
      });
    } catch (error) {
      throw error;
    }
  },
  async deleteSession(sessionID, userID, req, res) {
    try {
      if (!import_mongoose3.Types.ObjectId.isValid(sessionID)) {
        return res.status(400).jsend.fail({ code: 400, message: "Invalid ID format" });
      }
      const session = await Session.findOne({
        _id: sessionID,
        user: userID
      });
      if (!session) {
        return res.status(403).jsend.fail({ code: 403, message: "Action is not permitted" });
      }
      const refreshToken = req.cookies?.["refresh_token"];
      const now = Date.now();
      const createdTime = new Date(session.createdAt).getTime();
      const MIN_SESSION_ACTION_TIME = 36 * 60 * 60 * 1e3;
      if (session.refresh !== refreshToken && now - createdTime < MIN_SESSION_ACTION_TIME) {
        return res.status(403).jsend.fail({
          code: 403,
          message: "New users not allowed to delete other sessions"
        });
      }
      const result = await Session.deleteOne({ _id: sessionID });
      if (result.deletedCount === 0) {
        return res.status(500).jsend.success({ code: 500, message: "Internal server error" });
      }
      res.status(200).jsend.success({ code: 200, message: "Deleted successfully" });
    } catch (error) {
      throw error;
    }
  }
};

// app/routes/session/session.route.ts
var router = (0, import_express.Router)();
router.get(
  "/all",
  authenticateToken,
  async (req, res) => {
    try {
      await SessionService.getUserSessions(req.user?._id, res);
    } catch (error) {
      res.status(500).jsend.error({
        code: 500,
        message: "Something went wrong",
        data: error
      });
    }
  }
);
router.delete(
  "/delete/:id",
  authenticateToken,
  async (req, res, _next) => {
    try {
      await SessionService.deleteSession(req.params?.id, req.user?._id, req, res);
    } catch (error) {
      res.status(500).jsend.error({
        code: 500,
        message: "Something went wrong",
        data: error
      });
    }
  }
);
var session_route_default = router;